chore(helper-plugin)!: move useFetchClient and getFetchClient to the admin
