+++ packages/core/admin/admin/src/utils/baseQuery.ts
@@ -1,7 +1,8 @@
 import { SerializedError } from '@reduxjs/toolkit';
 import { BaseQueryFn } from '@reduxjs/toolkit/query';
-import { getFetchClient } from '@strapi/helper-plugin';
 import { isAxiosError, type AxiosRequestConfig } from 'axios';
+
+import { getFetchClient } from '../utils/getFetchClient';
 
 import type { ApiError } from '../types/errors';
