+++ packages/core/admin/admin/src/utils/tests/getFetchClient.test.ts
@@ -1,8 +1,9 @@
+import { auth } from '@strapi/helper-plugin';
 import { server } from '@tests/utils';
+import { AxiosError } from 'axios';
 import { rest } from 'msw';
 
-import { auth } from '../auth';
-import { instance } from '../fetchClient';
+import { getFetchClient, instance } from '../getFetchClient';
 
 const token = 'coolToken';
 const mockClearAppStorage = jest.fn().mockImplementation();
@@ -10,7 +11,7 @@ const mockClearAppStorage = jest.fn().mockImplementation();
 auth.getToken = jest.fn().mockReturnValue(token);
 auth.clearAppStorage = mockClearAppStorage;
 
-describe('HELPER-PLUGIN | utils | fetchClient', () => {
+describe('fetchClient', () => {
   it('should add the authorization token in each request', async () => {
     const response = await instance.get('/test-fetch-client');
     expect(response.config.headers.Authorization).toBe(`Bearer ${token}`);
@@ -62,6 +63,44 @@ describe('HELPER-PLUGIN | utils | fetchClient', () => {
     } catch (error) {
       expect(mockClearAppStorage).toHaveBeenCalled();
       expect(mockReload).toHaveBeenCalled();
+    }
+
+    server.restoreHandlers();
+  });
+});
+
+describe('getFetchClient', () => {
+  it('should return the 4 HTTP methods to call GET, POST, PUT and DELETE apis', () => {
+    const response = getFetchClient();
+    expect(response).toHaveProperty('get');
+    expect(response).toHaveProperty('post');
+    expect(response).toHaveProperty('put');
+    expect(response).toHaveProperty('del');
+  });
+
+  it('should contain the normalized URL when we try to reach an unknown API with an URL without prepending slash', async () => {
+    const response = getFetchClient();
+    try {
+      await response.get('test-fetch-client');
+    } catch (err) {
+      const url = (err as AxiosError).config?.url;
+      expect(url).toBe('/test-fetch-client');
+    }
+  });
+
+  it('should contain the URL passed when we try to reach an unknown API with a full URL', async () => {
+    server.use(
+      rest.get('https://notinternalurl.com', (req, res, ctx) => {
+        return res(ctx.status(200));
+      })
+    );
+
+    const response = getFetchClient();
+    try {
+      await response.get('https://notinternalurl.com');
+    } catch (err) {
+      const url = (err as AxiosError).config?.url;
+      expect(url).toBe('https://notinternalurl.com');
     }
 
     server.restoreHandlers();
