+++ packages/core/admin/admin/tests/server.ts
@@ -1223,6 +1223,27 @@ export const server = setupServer(
       );
     }),
     /**
+     *
+     * fetchClient, useFetchClient and getFetchClient
+     *
+     */
+    rest.get('/use-fetch-client-test', (req, res, ctx) => {
+      return res(
+        ctx.json({
+          data: {
+            results: [
+              { id: 2, name: 'newest', publishedAt: null },
+              { id: 1, name: 'oldest', publishedAt: null },
+            ],
+            pagination: { page: 1, pageCount: 10 },
+          },
+        })
+      );
+    }),
+    rest.get('/test-fetch-client', (req, res, ctx) => {
+      return res(ctx.status(200));
+    }),
+    /**
      * Content History
      */
     ...historyHandlers,
