+++ packages/core/helper-plugin/MIGRATION_GUIDE.md
@@ -528,6 +528,18 @@ import { useCollator } from '@strapi/helper-plugin';
 import { useCollator } from '@strapi/design-system';
 ```
 
+### useFetchClient
+
+This hook has been removed. You should import it from the `@strapi/strapi/admin` package:
+
+```tsx
+// Before
+import { useFetchClient } from '@strapi/helper-plugin';
+
+// After
+import { useFetchClient } from '@strapi/strapi/admin';
+```
+
 ### useFilter
 
 This component has been removed. You should import it from the `@strapi/design-system` package:
@@ -563,6 +575,18 @@ This util has been removed and not replaced. If you feel like you need this util
 ### getAPIInnerErrors
 
 This util has been removed and not replaced, use async / await with try / catch instead. If you feel like you need this util, please open an issue on the Strapi repository to discuss your usecase.
+
+### getFetchClient
+
+This util has been removed. You should import it from the `@strapi/strapi/admin` package:
+
+```tsx
+// Before
+import { getFetchClient } from '@strapi/helper-plugin';
+
+// After
+import { getFetchClient } from '@strapi/strapi/admin';
+```
 
 ### getFileExtension
