--- packages/core/helper-plugin/src/hooks/useFetchClient.ts
@@ -1,32 +1,0 @@
-/**
- *
- * useFetchClient
- *
- */
-import { useEffect, useMemo, useRef } from 'react';
-
-import { getFetchClient } from '../utils/getFetchClient';
-
-const useFetchClient = () => {
-  const controller = useRef<AbortController | null>(null);
-
-  if (controller.current === null) {
-    controller.current = new AbortController();
-  }
-
-  useEffect(() => {
-    return () => {
-      controller.current!.abort();
-    };
-  }, []);
-
-  return useMemo(
-    () =>
-      getFetchClient({
-        signal: controller.current!.signal,
-      }),
-    []
-  );
-};
-
-export { useFetchClient };
