+++ packages/core/helper-plugin/src/index.ts
@@ -11,7 +11,6 @@ export * from './features/RBAC';
  * -----------------------------------------------------------------------------------------------*/
 
 export * from './hooks/useClipboard';
-export * from './hooks/useFetchClient';
 export * from './hooks/useFocusInputField';
 export * from './hooks/useFocusWhenNavigate';
 export * from './hooks/useLockScroll';
@@ -26,7 +25,6 @@ export * from './hooks/useSelectionState';
  * -----------------------------------------------------------------------------------------------*/
 
 export * from './utils/auth';
-export * from './utils/getFetchClient';
 export * from './utils/getYupInnerErrors';
 export * from './utils/hasPermissions';
 export * from './utils/prefixPluginTranslations';
