+++ packages/core/upload/admin/src/components/EditFolderDialog/tests/EditFolderDialog.test.jsx
@@ -14,6 +14,10 @@ import { EditFolderDialog } from '../EditFolderDialog';
 jest.mock('@strapi/helper-plugin', () => ({
   ...jest.requireActual('@strapi/helper-plugin'),
   useQueryParams: jest.fn().mockReturnValue([{ query: {} }]),
+}));
+
+jest.mock('@strapi/admin/strapi-admin', () => ({
+  ...jest.requireActual('@strapi/admin/strapi-admin'),
   useFetchClient: jest.fn().mockReturnValue({
     put: jest.fn().mockImplementation({}),
   }),
