+++ packages/core/upload/admin/src/hooks/useBulkRemove.js
@@ -1,4 +1,5 @@
-import { useFetchClient, useNotification } from '@strapi/helper-plugin';
+import { useFetchClient } from '@strapi/admin/strapi-admin';
+import { useNotification } from '@strapi/helper-plugin';
 import { useMutation, useQueryClient } from 'react-query';
 
 import pluginId from '../pluginId';
