+++ packages/core/upload/admin/src/hooks/useFolderStructure.js
@@ -1,4 +1,4 @@
-import { useFetchClient } from '@strapi/helper-plugin';
+import { useFetchClient } from '@strapi/admin/strapi-admin';
 import { useIntl } from 'react-intl';
 import { useQuery } from 'react-query';
