+++ packages/plugins/i18n/admin/src/utils/baseQuery.ts
@@ -1,6 +1,6 @@
 import { SerializedError } from '@reduxjs/toolkit';
 import { BaseQueryFn } from '@reduxjs/toolkit/query';
-import { getFetchClient } from '@strapi/helper-plugin';
+import { getFetchClient } from '@strapi/strapi/admin';
 import { isAxiosError, type AxiosRequestConfig } from 'axios';
 
 import type { ApiError } from '@strapi/strapi/admin';
