+++ packages/plugins/users-permissions/admin/src/pages/AdvancedSettings/index.jsx
@@ -13,13 +13,12 @@ import {
   useNotifyAT,
 } from '@strapi/design-system';
 import {
-  useFetchClient,
   useFocusWhenNavigate,
   useNotification,
   useRBAC,
 } from '@strapi/helper-plugin';
 import { Check } from '@strapi/icons';
-import { useAPIErrorHandler, Page, Form, InputRenderer } from '@strapi/strapi/admin';
+import { useAPIErrorHandler, useFetchClient, Page, Form, InputRenderer } from '@strapi/strapi/admin';
 import { Helmet } from 'react-helmet';
 import { useIntl } from 'react-intl';
 import { useMutation, useQuery, useQueryClient } from 'react-query';
