+++ packages/plugins/users-permissions/admin/src/pages/EmailTemplates/index.jsx
@@ -3,12 +3,11 @@ import * as React from 'react';
 import { useTracking } from '@strapi/admin/strapi-admin';
 import { ContentLayout, HeaderLayout, Main, useNotifyAT } from '@strapi/design-system';
 import {
-  useFetchClient,
   useFocusWhenNavigate,
   useNotification,
   useRBAC,
 } from '@strapi/helper-plugin';
-import { Page, useAPIErrorHandler } from '@strapi/strapi/admin';
+import { Page, useAPIErrorHandler, useFetchClient } from '@strapi/strapi/admin';
 import { Helmet } from 'react-helmet';
 import { useIntl } from 'react-intl';
 import { useMutation, useQuery, useQueryClient } from 'react-query';
