+++ packages/plugins/users-permissions/admin/src/pages/Providers/index.jsx
@@ -20,13 +20,12 @@ import {
 import {
   onRowClick,
   stopPropagation,
-  useFetchClient,
   useFocusWhenNavigate,
   useNotification,
   useRBAC,
 } from '@strapi/helper-plugin';
 import { Pencil } from '@strapi/icons';
-import { Page, useAPIErrorHandler } from '@strapi/strapi/admin';
+import { Page, useAPIErrorHandler, useFetchClient } from '@strapi/strapi/admin';
 import upperFirst from 'lodash/upperFirst';
 import { Helmet } from 'react-helmet';
 import { useIntl } from 'react-intl';
