+++ packages/plugins/users-permissions/admin/src/pages/Roles/pages/EditPage.jsx
@@ -12,9 +12,9 @@ import {
   GridItem,
   Grid,
 } from '@strapi/design-system';
-import { useFetchClient, useNotification } from '@strapi/helper-plugin';
+import { useNotification } from '@strapi/helper-plugin';
 import { Check } from '@strapi/icons';
-import { Page, BackButton, useAPIErrorHandler } from '@strapi/strapi/admin';
+import { Page, BackButton, useAPIErrorHandler, useFetchClient } from '@strapi/strapi/admin';
 import { Formik, Form } from 'formik';
 import { Helmet } from 'react-helmet';
 import { useIntl } from 'react-intl';
